Add links on movie banner
