+++ apps/festival/festival/src/app/dashboard/title/details/details.component.html
@@ -89,15 +89,3 @@
     <movie-summary-file [promotionalElements]="form.promotionalElements" [link]="getPath('files&links')"></movie-summary-file>
   </mat-card-content>
 </mat-card>
-
-<!-- Evaluation card -->
-<mat-card>
-  <mat-card-header fxLayout="row" fxLayoutAlign="start center">
-    <mat-card-title>Marketplace Evaluation</mat-card-title>
-        <!-- TODO #2570 -->
-    <!-- <a mat-icon-button [routerLink]="getPath('evaluation')"><mat-icon svgIcon="pencil"></mat-icon></a> -->
-  </mat-card-header>
-  <mat-card-content>
-    <movie-summary-evaluation [salesInfo]="form.salesInfo" [link]="getPath('evaluation')"></movie-summary-evaluation>
-  </mat-card-content>
-</mat-card>
