+++ apps/festival/festival/src/app/marketplace/title/view/view.component.html
@@ -8,14 +8,18 @@
             <span>Contact Sales Agent</span>
             <mat-icon svgIcon="arrow_forward"></mat-icon>
           </a>
+          <article *ngIf="hasLink(movie)" fxLayout="row" fxLayout.lt-md="column" fxLayoutGap="24px">
+            <wishlist-button [movieId]="movie.id" small>
+            </wishlist-button>
+            <ng-container *ngFor="let promoLink of promoLinks">
+              <a *ngIf="getLink(movie, promoLink) as link" [href]="link.url" target="_blank" fxLayout="row" fxLayoutAlign="start center">
+                <mat-icon [svgIcon]="link.icon"></mat-icon>
+                <span>{{ link.label }}</span>
+              </a>
+            </ng-container>
+          </article>
         </ng-container>
       </movie-header-cta>
-      <movie-slider-actions>
-        <wishlist-button [movieId]="movie.id" small>
-          <wishlist-add-text>Add to Wishlist</wishlist-add-text>
-          <wishlist-remove-text>Remove from Wishlist</wishlist-remove-text>
-        </wishlist-button>
-      </movie-slider-actions>
     </movie-header>
   </movie-view>
 </ng-container>
