+++ apps/festival/festival/src/app/marketplace/title/view/view.component.ts
@@ -1,16 +1,15 @@
 import { Component, OnInit, ChangeDetectionStrategy, HostBinding } from '@angular/core';
 import { Router } from '@angular/router';
-import { MatSnackBar } from '@angular/material/snack-bar';
-import { CartService } from '@blockframes/cart/+state/cart.service';
-import { FireAnalytics } from '@blockframes/utils/analytics/app-analytics';
 import { Observable } from 'rxjs';
-import { map, switchMap } from 'rxjs/operators';
+import { switchMap } from 'rxjs/operators';
 import { Movie } from '@blockframes/movie/+state/movie.model';
 import { MovieQuery } from '@blockframes/movie/+state/movie.query';
 import { Organization } from '@blockframes/organization/+state/organization.model';
 import { OrganizationService } from '@blockframes/organization/+state/organization.service';
-import { OrganizationQuery } from '@blockframes/organization/+state/organization.query';
 import { scaleIn } from '@blockframes/utils/animations/fade';
+import { getLabelBySlug, ExtractCode } from '@blockframes/utils/static-model/staticModels';
+
+const promoLinks = ['promo_reel_link', 'scenario', 'screener_link', 'teaser_link', 'presentation_deck', 'trailer_link'];
 
 @Component({
   selector: 'festival-movie-view',
@@ -23,9 +22,7 @@ export class ViewComponent implements OnInit {
   @HostBinding('@scaleIn') animPage;
   public movie$: Observable<Movie>;
   public orgs$: Observable<Organization[]>;
-  // Flag to indicate which icon and message to show
-  public toggle$: Observable<boolean>;
-  public events = [ new Date(), new Date() ];
+  public promoLinks = promoLinks;
 
   navLinks = [{
     path: 'main',
@@ -34,11 +31,7 @@ export class ViewComponent implements OnInit {
 
   constructor(
     private movieQuery: MovieQuery,
-    private cartService: CartService,
     private orgService: OrganizationService,
-    private orgQuery: OrganizationQuery,
-    private snackbar: MatSnackBar,
-    private analytics: FireAnalytics,
     public router: Router
   ) {}
 
@@ -47,34 +40,20 @@ export class ViewComponent implements OnInit {
     this.orgs$ = this.movieQuery.selectActiveId().pipe(
       switchMap(movieId => this.orgService.getValue(ref => ref.where('movieIds', 'array-contains', movieId)))
     );
-    this.toggle$ = this.orgQuery.selectActive().pipe(
-      map(org => {
-        return org.wishlist
-          .filter(({ status }) => status === 'pending')
-          .some(({ movieIds }) => movieIds.includes(this.movieQuery.getActiveId()));
-      })
-    );
   }
 
-  public addToWishlist() {
-    const movie = this.movieQuery.getActive();
-    const title = movie.main.title.international;
-    this.cartService.updateWishlist(movie);
-    this.analytics.event('addedToWishlist', {
-      movieId: movie.id,
-      movieTitle: movie.main.title.original
-    });
-    this.snackbar.open(`${title} has been added to your selection.`, 'close', { duration: 2000 });
+  public hasLink({ promotionalElements }: Movie): boolean {
+    return this.promoLinks.some(link => !!promotionalElements[link].media.url);
   }
 
-  public removeFromWishlist() {
-    const movie = this.movieQuery.getActive();
-    const title = movie.main.title.international;
-    this.cartService.updateWishlist(movie);
-    this.analytics.event('removedFromWishlist', {
-      movieId: movie.id,
-      movieTitle: movie.main.title.original
-    });
-    this.snackbar.open(`${title} has been removed from your selection.`, 'close', { duration: 2000 });
+  public getLink(movie: Movie, link: ExtractCode<'PROMOTIONAL_ELEMENT_TYPES'>) {
+    if(movie.promotionalElements[link].media.url) {
+      const isDownload = link === 'scenario' || link === 'presentation_deck';
+      return {
+        url: movie.promotionalElements[link].media.url,
+        icon: isDownload ? 'download' : 'play',
+        label: isDownload ? `Download ${getLabelBySlug('PROMOTIONAL_ELEMENT_TYPES', link)}` : `Watch ${getLabelBySlug('PROMOTIONAL_ELEMENT_TYPES', link)}`
+      };
+    }
   }
 }
