+++ libs/movie/src/lib/movie/components/slider/slider.module.ts
